+++ src/midilfo_lv2.h
@@ -99,7 +99,7 @@ private:
         float *val[35];
         uint64_t curFrame;
         uint64_t tempoChangeTick;
-        int curTick;
+        uint64_t curTick;
         int inLfoFrame;
         double mouseXCur;
         double mouseYCur;
