+++ src/midiseq.h
@@ -185,7 +185,7 @@ class MidiSeq : public MidiWorker  {
  * @param tick the current tick at which we request a note. This tick will be
  * used to calculate the nextTick which is quantized to the pattern
  */
-    void getNextFrame(int tick);
+    void getNextFrame(int64_t tick);
 /*! @brief  toggles the mute state of one point of the
  * MidiSeq::muteMask array.
  *
